refactor(viewer): modify useDrawingHandler hook import location
